% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqfexplore.R
\name{dqf.explore}
\alias{dqf.explore}
\title{Interactive Visualization of DQF information}
\usage{
dqf.explore(dqfs, show = c())
}
\arguments{
\item{dqfs}{a dqf.outlier object}

\item{show}{vector of preselected observation indicies to highlight}
}
\description{
Creates 3 interactive plots of DQFs with summarization on a 4th
}
\examples{
fit.dqf <- dqf.outlier(iris[51:102,1:4], g.scale=6)
dqf.explore(fit.dqf)
}
\keyword{dqf}
