% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plothelp.R
\name{plot.dqf}
\alias{plot.dqf}
\title{Visualizes DQFs computed from dqf.outlier function}
\usage{
\method{plot}{dqf}(dqf, angle, click.points)
}
\arguments{
\item{dqf}{a matrix of dqfs}

\item{angle}{angle used in computation of dqf}

\item{click.points}{vector of points to highlight}
}
\description{
Dependency function for dqf.explore
}
\keyword{dqf}
