% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwin.R
\name{sd.w}
\alias{sd.w}
\title{Windosrized Standard Deviation Computation}
\usage{
sd.w(x, k)
}
\arguments{
\item{x}{univariate data vector}

\item{k}{integer number of points to be windorized from each extreme}
}
\description{
Computes the windorized standard deviation of a univariate data set
}
\examples{
sd.w(rnorm(100), 3)
}
\keyword{dqf}
