% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsample.R
\name{subsamp.dqf}
\alias{subsamp.dqf}
\title{Balanced subsample of all pairwise comparisons}
\usage{
subsamp.dqf(n.obs, subsample)
}
\arguments{
\item{subsample}{number of pairs each observation shows up in}

\item{n.osb}{integer number of data point}
}
\description{
To reduce computational complexity of DQF method
}
\examples{
subsamp.dqf(5,2)
}
\keyword{dqf}
